API documentation

Orthographic system

Lycian.LycianAscii — Type

An orthographic system for encoding the Lycian alphabet primarily using ASCII characaters.

Lycian.lycianAscii — Function

Instantiate a LycianAscii with correct code points and token types.

lycianAscii()

Orthography.codepoints — Function

Delegate to specific functions based on type's orthography trait value.

codepoints(x)

It is an error to invoke the codepoints function on anything but an orthographic system.

codepoints(_, x)

Orthographic systems must implement codepoints.

codepoints(_, ortho)

Implement codepoints function for SimpleAscii.

codepoints(ortho)

Implement Orthography's codepoints functions for LycianAscii.

codepoints(ortho)

Orthography.tokentypes — Function

Delegate to specific functions based on type's orthography trait value.

It is an error to invoke the tokentypes function on anything but an orthographic system.

tokentypes(_, x)

Orthographic systems must implement tokentypes.

tokentypes(_, ortho, s)

Implement tokentypes function for SimpleAscii.

tokentypes(ortho)

Implement Orthography's tokentypes functions for LycianAscii.

tokentypes(ortho)

CitableParserBuilder.parsetoken — Function

Delegate to specific functions based on type's citable trait value.

parsetoken(s, x; data)

It is an error to invoke the parsetoken using types that are not a parser.

parsetoken(, s, x; data)

Citable parsers must implement parsetoken.

parsetoken(, s, x; data)

Parse String s by looking it up in a given dictionary.

Parse a single token in LycianAscii orthography.

parsetoken(s, parser)
parsetoken(s, parser, data)

Lycian.ucode — Function

Convert ASCII encoding to Lycian Unicode range.

ucode(s)

Internal functions:

Lycian.tokenforstring — Function

Create correct OrthographicToken for a given string.

tokenforstring(s)

Lycian.alphabetic — Function

Compose a string with all alphabetic characters.

alphabetic()

Lycian.isNumeric — Function

True if all characters in s are numeric.

isNumeric(s)

Lycian.isAlphabetic — Function

True if all characters in s are alphabetic.

isAlphabetic(s)

Lycian.isPunctuation — Function

True if s is the interpunctuation mark.

isPunctuation(s)

Lycian.a2uDict — Function

Compose a map of ASCIIish characters to Lycian unicode code points.

a2uDict()

Lexicon

Lycian.lexicon_df — Function

Construct a DataFrame with Lycian lexicon.

lexicon_df()

Lycian.lookuplex — Function

Lookup entry identified by LexemeUrn in Lycian lexicon.

lookuplex(lexicon, lexu)

Lookup entry identified by Cite2Urn in Lycian lexicon.

lookuplex(lexicon, u)

Lookup entry identified by URN string in Lycian lexicon.

lookuplex(lexicon, urnstring)

Lycian.searchlex — Function

Search lexicon for entries where transliterated lemma matches s

searchlex(lexicon, s)

Morphology

Lycian.morph_df — Function

Construct a DataFrame with Lycian morphology.

morph_df()

Lycian.searchmorph — Function

Search morphological data for forms matchng s

searchmorph(morphdf, s)

Lycian.forms_df — Function

Construct a DataFrame with Lycian morphology.

forms_df()

Lycian.lookupform — Function

Lookup entry identified by FormUrn in DataFrame of Lycian forms.

lookupform(formsdf, formu)

Lookup entry identified by Cite2Urn in DataFrame of Lycian forms..

lookupform(formsdf, u)

Lookup entry identified by URN string in DataFrame of Lycian forms..

lookupform(formsdf, urnstring)